1youtube_short_script_assistant/youtube_short_script_assistant/pyproject.toml
[project]
name = "adkllmproject"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "deprecated>=1.2.18",
    "google-adk==0.1.0",
]
